Add files via upload
It's my chance to pursue certain things that we say all the time and fear making them so true. And somehow I made it possible for the first time with the help of a community of developers. I initiated this startup with the virtue of doing something away from the certain line of work I'm going to do for my lifetime. So going on with a pinch of motivation, let's see where it can go to an extent.